+++ src/rex.deploy/src/htsql_rex_deploy/tr/__init__.py
@@ -3,6 +3,6 @@
 #
 
 
-from . import bind, dump, encode, lookup, signature
+from . import bind, dump, encode, lookup, reduce, signature